server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.5
frozenlist==1.3.3
hiredis==2.2.2
idna==3.4
inflection==0.5.1
internetarchive==3.4.0
itypes==1.2.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
psycopg==3.1.8
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schema==0.7.5
simplejson==3.19.1
sqlparse==0.4.3
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
